common/xfs: Execute _xfs_check only when explicitly asked
fsstress when executed as part of some of the tests (e.g. generic/270)
invokes chown() syscall many times by passing random integers as value
for the uid argument. For each such syscall invocation for which there
is no on-disk quota block, xfs invokes xfs_dquot_disk_alloc() which
allocates a new block and instantiates all the quota structures mapped
by the newly allocated block. For filesystems with 64k block size, the
number of on-disk quota structures created will be 16 times more than
that for a filesystem with 4k block size.

xfs_db's check command (executed after test script finishes execution)
will try to read in all of the on-disk quota structures into
memory. This causes the OOM event to be triggered when reading from
filesystems with 64k block size. For filesystems with sufficiently large
amount of system memory, this causes the test to execute for a very long
time.

Due to the above stated reasons, this commit disables execution of
xfs_db's check command unless explictly asked by the user by setting
$EXECUTE_XFS_DB_CHECK variable.

Signed-off-by: Chandan Rajendra <chandanrlinux@gmail.com>